rename props functions
